+++ scarab-engine/src/gameobject/entity/effect_helpers.rs
@@ -1,13 +1,13 @@
 use graphics::types::Scalar;
 use serde::{Deserialize, Serialize};
 
+use super::registry::RegisteredEntity;
 use crate::{
     gameobject::HasHealth,
     scene::{PendingEffect, TargetsOthers},
-    PhysBox, ScarabResult,
+    types::physbox::PhysBox,
+    ScarabResult,
 };
-
-use super::registry::RegisteredEntity;
 
 #[derive(Debug, Serialize, Deserialize, Default, Clone, Copy, PartialEq)]
 /// Expresses the readiness of an effect
